fix: use double for triangle elevation

+++ lib/src/indicators/triangle_indicator.dart
@@ -4,7 +4,7 @@ class TriangleIndicator extends StatelessWidget {
   final Color? color;
   final double width;
   final double height;
-  final int elevation;
+  final double elevation;
 
   const TriangleIndicator({
     Key? key,
